requirements/base.txt
django~=3.2
djangorestframework == 3.14.0
celery == 5.2.7
tenant-schemas-celery == 2.0.0
django-tenants~=3.3  # Use multitenants - to get multiple periods of activities

djangorestframework-datatables==0.5.1  # Pagination for datatables. DO NOT UPOGRADE TO 0.7.0 unless django is migrated to 4.0
django-dbtemplates == 4.0  # Store overrides of templates in db, edit through admin
django-localflavor == 3.1
django-model-utils == 4.3.1  # Gewneral model toolkit
django-sekizai == 4.0.0  # JS and CSS as blocks
django-countries == 7.5  # country field
django-formtools == 2.4  # form wizard
django-anymail == 8.6  # email backend
django-dynamic-preferences == 1.14.0  # Settings that can be edited in admin
python-dateutil == 2.8.2  # date parsing
django-registration-redux == 2.11  # Registration view
django-recaptcha==3.0.0
django-crispy-forms==1.14.0
django-ckeditor == 6.5.1
Pillow == 9.3.0
django-phonenumber-field == 7.0.1
phonenumbers == 8.13.2

# Excel export
django-import-export == 3.0.2

django-floppyforms==1.9.0
django-braces == 1.14.0
django-select2 ~= 8.1.0
django-autoslug == 1.9.8
# WYSIWYG editor
django-picklefield == 1.0.0
django-multiupload == 0.6.1
logutils==0.3.5
# skin for admin
django-grappelli == 2.8.1
# addtoblock template tag (js embedding)
# xlrd == 1.1.0
pypdftk==0.5
django-mathfilters==0.4.0
django-async-messages == 0.3.1


django_environ == 0.4.4
requests==2.28.1
python-memcached==1.59
